refactor: 전화번호 중복 사장님으로 이관
