+++ src/main/java/in/koreatech/koin/domain/owner/controller/OwnerController.java
@@ -27,6 +27,7 @@ import in.koreatech.koin.domain.owner.dto.OwnerSendEmailRequest;
 import in.koreatech.koin.domain.owner.dto.OwnerSendSmsRequest;
 import in.koreatech.koin.domain.owner.dto.OwnerSmsVerifyRequest;
 import in.koreatech.koin.domain.owner.dto.OwnerVerifyResponse;
+import in.koreatech.koin.domain.owner.dto.PhoneNumberCheckExistsRequest;
 import in.koreatech.koin.domain.owner.dto.VerifyEmailRequest;
 import in.koreatech.koin.domain.owner.dto.VerifySmsRequest;
 import in.koreatech.koin.domain.owner.service.OwnerService;
@@ -159,6 +160,15 @@ public class OwnerController implements OwnerApi {
         @Valid CompanyNumberCheckRequest request
     ) {
         ownerService.checkCompanyNumber(request);
+        return ResponseEntity.ok().build();
+    }
+    
+    @GetMapping("/owners/check/phone-number")
+    public ResponseEntity<Void> checkDuplicationOfPhoneNumber(
+        @ModelAttribute("phone_number")
+        @Valid PhoneNumberCheckExistsRequest request
+    ) {
+        ownerService.checkExistsAccount(request);
         return ResponseEntity.ok().build();
     }
 }
